atualização: crud rodando redondo...

+++ RepositoryAPW/pages/controle-de-usuarios.aspx.cs
@@ -13,12 +13,11 @@ namespace RepositoryAPW.pages
     {
         private readonly webservicewithrest ws = new webservicewithrest();
         private readonly JavaScriptSerializer jss = new JavaScriptSerializer();
+        private readonly string url = "http://localhost:21315/pages/controle-de-usuarios.aspx";
+
         protected void Page_Load(object sender, EventArgs e)
         {
-            if (!IsPostBack)
-            {
-                gvDataBind();
-            }
+            gvDataBind();
         }
 
         private void gvDataBind()
@@ -36,22 +35,46 @@ namespace RepositoryAPW.pages
             gvDataBind();
         }
 
-
         protected void btnGravar_Click(object sender, EventArgs e)
         {
-            if (Session["Id"] == null)
-            {
-                ws.create_user(tbNome.Text, tbEmail.Text, tbSenha.Text);
-                gvDataBind();
+            var id = Session["Id"];
 
-                Auxiliar.ClearControls(this);
+            try
+            {
+                if (id == null)
+                {
+                    if (!ws.check_user(tbEmail.Text))
+                    {
+                        ws.create_user(tbNome.Text, tbEmail.Text, tbSenha.Text);
+                        gvDataBind();
+
+                        Auxiliar.ClearControls(this);
+                    }
+                    else
+                    {
+                        throw new Exception("Já existe cadastro com este email.");
+                    }
+                }
+                else
+                {
+                    ws.update_user(Convert.ToInt32(id), tbNome.Text, tbEmail.Text, tbSenha.Text);
+                    gvDataBind();
+
+                    tbEmail.Enabled = true;
+
+                    Auxiliar.ClearControls(this);
+
+                    throw new Exception("Dados atualizados com sucesso.");
+                }
+
+                Session.Clear();
+
             }
-            else
+            catch (Exception ex)
             {
-                ws.update_user(Convert.ToInt32(Session["Id"]), tbNome.Text, tbEmail.Text, tbSenha.Text);
-                gvDataBind();
-
-                Auxiliar.ClearControls(this);
+                lblMensagem.Visible = true;
+                lblMensagem.Text = ex.Message;
+                Session.Clear();
             }
         }
 
@@ -61,43 +84,39 @@ namespace RepositoryAPW.pages
             {
                 var id = Convert.ToInt32(e.CommandArgument);
 
-                switch (e.CommandName)
+                if (e.CommandName == "Select")
                 {
-                    case "add":
+                    tbEmail.Enabled = false;
 
-                        var dado = ws.return_user_by_id(id.ToString());
+                    var dado = ws.return_user_by_id(id.ToString());
 
-                        var usuario = jss.Deserialize<Usuario[]>(dado);
+                    var usuario = jss.Deserialize<Usuario[]>(dado);
 
-                        Session["Id"] = usuario[0].Id;
-                        tbNome.Text = usuario[0].Nome;
-                        tbEmail.Text = usuario[0].Email;
-                        tbSenha.Text = usuario[0].Senha;
-
-                        break;
-
-                    case "delete":
-
-                        ws.delete_user(id);
-
-                        gvDataBind();
-
-                        break;
-
-                    default:
-                        break;
+                    Session["Id"] = usuario[0].Id;
+                    tbNome.Text = usuario[0].Nome;
+                    tbEmail.Text = usuario[0].Email;
+                    tbSenha.Text = usuario[0].Senha;
+                }
+                else if (e.CommandName == "delete")
+                {
+                    ws.delete_user(id);
+                    gvDataBind();
+                    Response.Redirect(url);
                 }
             }
             catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                lblMensagem.Visible = true;
+                lblMensagem.Text = ex.Message;
             }
         }
 
         protected void btnLimpar_Click(object sender, EventArgs e)
         {
             Auxiliar.ClearControls(this);
-            gvDataBind();
+            lblMensagem.Text = "";
+            tbEmail.Enabled = true;
+            Session.Clear();
         }
     }
 }
